fix: 신청페이지 - 신청, 내역 확인 버튼 수정
