+++ src/styles/applyEdu.css
@@ -1,21 +1,21 @@
 .typeSelect-container{
-    margin-left: auto;
-    margin-right: auto;
-    width: 65.571rem;
-    flex-grow: 0;
-    margin-top: 4.677rem;
+    display: grid;
+    width: 65rem;
+    grid-template-columns: 1fr 1fr;
+    gap:1.111rem;
+    /* margin-top: 4.677rem; */
     margin-bottom: 4.677rem;
+
 }
 
 .typeSelect-button{
-    width: 25rem;
     background: #F8F9FD;
     border-radius: 10px;
     font-family: 'Pretendard';
-    font-style: normal;
+    font-style: bold;
     font-weight: 700;
     font-size: 2rem;
     color: #0072B9;
-    margin: 2rem;
-
+    height: 6.111rem;
+    font-size: 2.667rem;
 }
