+++ src/styles/makeNewApplication.css
@@ -155,7 +155,6 @@
 
 .CreateEdu-content-root {
   width: 65.571rem;
-  flex-grow: 0;
   margin: 4.444rem auto 7.677rem auto;
 }
